# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  service.JiraComponent:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  service.JiraIssue:
    properties:
      fields:
        $ref: '#/definitions/service.JiraIssueFields'
      id:
        type: string
      key:
        type: string
      link:
        type: string
      project:
        type: string
    type: object
  service.JiraIssueFields:
    properties:
      assignee:
        $ref: '#/definitions/service.JiraUser'
      components:
        items:
          $ref: '#/definitions/service.JiraComponent'
        type: array
      created:
        type: string
      description:
        type: string
      issuetype:
        $ref: '#/definitions/service.JiraIssueType'
      labels:
        items:
          type: string
        type: array
      parent:
        $ref: '#/definitions/service.JiraParent'
      priority:
        $ref: '#/definitions/service.JiraPriority'
      reporter:
        $ref: '#/definitions/service.JiraUser'
      status:
        $ref: '#/definitions/service.JiraStatus'
      subtasks:
        items:
          $ref: '#/definitions/service.JiraSubtask'
        type: array
      summary:
        type: string
      updated:
        type: string
    type: object
  service.JiraIssueType:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  service.JiraIssuesResponse:
    properties:
      issues:
        items:
          $ref: '#/definitions/service.JiraIssue'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  service.JiraParent:
    properties:
      fields:
        $ref: '#/definitions/service.JiraParentFields'
      id:
        type: string
      key:
        type: string
    type: object
  service.JiraParentFields:
    properties:
      issuetype:
        $ref: '#/definitions/service.JiraIssueType'
      priority:
        $ref: '#/definitions/service.JiraPriority'
      status:
        $ref: '#/definitions/service.JiraStatus'
      summary:
        type: string
    type: object
  service.JiraPriority:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  service.JiraStatus:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  service.JiraSubtask:
    properties:
      fields:
        $ref: '#/definitions/service.JiraSubtaskFields'
      id:
        type: string
      key:
        type: string
    type: object
  service.JiraSubtaskFields:
    properties:
      issuetype:
        $ref: '#/definitions/service.JiraIssueType'
      priority:
        $ref: '#/definitions/service.JiraPriority'
      status:
        $ref: '#/definitions/service.JiraStatus'
      summary:
        type: string
    type: object
  service.JiraUser:
    properties:
      accountId:
        type: string
      displayName:
        type: string
      emailAddress:
        type: string
    type: object
  service.LDAPUserSearchItem:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      mobile:
        type: string
      new:
        type: boolean
    type: object
  service.LDAPUserSearchResponse:
    properties:
      result:
        items:
          $ref: '#/definitions/service.LDAPUserSearchItem'
        type: array
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
...
  /cis-public/proxy:
    get:
      consumes:
      - application/json
      description: Fetch data from component public endpoints (health, version, system
        info) server-side to avoid CORS
      parameters:
      - description: Component public endpoint URL to fetch
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response from component endpoint
          schema:
            additionalProperties: true
            type: object
        "400":
          description: URL parameter is required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Failed to fetch from component
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Proxy component public endpoints
      tags:
      - cis-public
...
  /jira/issues:
    get:
      description: Returns Jira issues filtered by project, status, team, assignee,
        type, summary, and key using real Jira values with pagination
      parameters:
      - description: Jira project key (e.g., SAPBTPCFS)
        in: query
        name: project
        type: string
      - description: Jira status values (e.g., 'Open,In Progress,Re Opened')
        in: query
        name: status
        type: string
      - description: Team name for filtering
        in: query
        name: team
        type: string
      - description: Assignee username for filtering
        in: query
        name: assignee
        type: string
      - description: Issue type (e.g., 'Bug,Task,Story')
        in: query
        name: type
        type: string
      - description: Free text search in summary
        in: query
        name: summary
        type: string
      - description: Specific issue key (e.g., 'BUG-1234')
        in: query
        name: key
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Issues
          schema:
            $ref: '#/definitions/service.JiraIssuesResponse'
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Jira request failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Jira issues for teams/projects
      tags:
      - jira
  /jira/issues/me:
    get:
      description: Returns Jira issues for the current authenticated user with optional
        filtering and pagination
      parameters:
      - description: Jira status values (e.g., 'Open,In Progress')
        in: query
        name: status
        type: string
      - description: Jira project key (e.g., SAPBTPCFS)
        in: query
        name: project
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Issues
          schema:
            $ref: '#/definitions/service.JiraIssuesResponse'
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Jira request failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my Jira issues
      tags:
      - jira
  /jira/issues/me/count:
    get:
      description: Returns the count of Jira issues for the current authenticated
        user filtered by status
      parameters:
      - description: Jira status value (e.g., 'Resolved')
        in: query
        name: status
        required: true
        type: string
      - description: Jira project key (e.g., SAPBTPCFS)
        in: query
        name: project
        type: string
      - description: 'Date in yyyy-MM-dd format for date filtering (default: one year
          ago for resolved issues)'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Count
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Missing or invalid query parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Jira request failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get count of my Jira issues by status
      tags:
      - jira
...
  /users/search/new:
    get:
      description: Searches LDAP directory for users where name starts with given
        prefix
      parameters:
      - description: Name prefix
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/service.LDAPUserSearchResponse'
        "400":
          description: Missing or invalid query parameter
          schema:
            additionalProperties: true
            type: object
        "502":
          description: LDAP connection or search failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search LDAP users by name prefix
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"